Refactor swipe listener code
