+++ app/src/main/java/com/github/h3lp3rs/h3lp/presentation/PresArrivalActivity.kt
@@ -1,4 +1,4 @@
-package com.github.h3lp3rs.h3lp
+package com.github.h3lp3rs.h3lp.presentation
 
 import android.annotation.SuppressLint
 import android.content.Intent
@@ -6,16 +6,17 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.GestureDetector
 import android.view.View
+import com.github.h3lp3rs.h3lp.R
 import com.github.h3lp3rs.h3lp.listeners.SwipeListener
 
 /**
  * Class representing the first page of the app presentation
  * Contains attractive commercial information
  */
-class PresentationActivity1 : AppCompatActivity() {
+class PresArrivalActivity : AppCompatActivity() {
     private val onSwipeLeft: () -> Unit = {
         // Launch next presentation page
-        val i = Intent(this, PresentationActivity2::class.java)
+        val i = Intent(this, PresRelevantActivity::class.java)
         startActivity(i)
         overridePendingTransition(R.anim.slide_in_right, R.anim.slide_out_left)
     }
