+++ app/src/main/java/com/github/h3lp3rs/h3lp/presentation/PresIrrelevantActivity.kt
@@ -1,4 +1,4 @@
-package com.github.h3lp3rs.h3lp
+package com.github.h3lp3rs.h3lp.presentation
 
 import android.annotation.SuppressLint
 import android.content.Intent
@@ -6,16 +6,18 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.GestureDetector
 import android.view.View
+import com.github.h3lp3rs.h3lp.MainPageActivity
+import com.github.h3lp3rs.h3lp.R
 import com.github.h3lp3rs.h3lp.listeners.SwipeListener
 
 /**
  * Class representing the third page of the app presentation
  * The purpose of this activity is to explain what H3LP ought not to be confused with
  */
-class PresentationActivity3 : AppCompatActivity() {
+class PresIrrelevantActivity : AppCompatActivity() {
     private val onSwipeRight: () -> Unit = {
         // Go to previous page
-        val i = Intent(this, PresentationActivity2::class.java)
+        val i = Intent(this, PresRelevantActivity::class.java)
         startActivity(i)
         overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right)
     }
