+++ app/src/main/java/com/github/h3lp3rs/h3lp/presentation/PresRelevantActivity.kt
@@ -1,4 +1,4 @@
-package com.github.h3lp3rs.h3lp
+package com.github.h3lp3rs.h3lp.presentation
 
 import android.annotation.SuppressLint
 import android.content.Intent
@@ -6,22 +6,23 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.GestureDetector
 import android.view.View
+import com.github.h3lp3rs.h3lp.R
 import com.github.h3lp3rs.h3lp.listeners.SwipeListener
 
 /**
  * Class representing the second page of the app presentation
  * The purpose of this activity is to explain what H3LP is about
  */
-class PresentationActivity2 : AppCompatActivity() {
+class PresRelevantActivity : AppCompatActivity() {
     private val onSwipeLeft: () -> Unit = {
         // Launch next presentation page
-        val i = Intent(this, PresentationActivity3::class.java)
+        val i = Intent(this, PresIrrelevantActivity::class.java)
         startActivity(i)
         overridePendingTransition(R.anim.slide_in_right, R.anim.slide_out_left)
     }
     private val onSwipeRight: () -> Unit = {
         // Go to previous page
-        val i = Intent(this, PresentationActivity1::class.java)
+        val i = Intent(this, PresArrivalActivity::class.java)
         startActivity(i)
         overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right)
     }
