+++ app/src/main/res/layout/activity_presentation2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PresentationActivity2">
+    tools:context=".presentation.PresRelevantActivity">
 
     <TextView
         android:id="@+id/pres2_textView6"
