+++ app/src/main/res/layout/activity_presentation3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PresentationActivity3">
+    tools:context=".presentation.PresIrrelevantActivity">
 
     <TextView
         android:id="@+id/pres3_textView2"
